Reintroduced shell script as compilation mechanism
Instead of Makefiles McSync is now built using a shell script. This removed the
"gnu make" dependency and is easily maintainable with the current McSync source
structure. The 'mcsync.sh' script takes care of rebuilding McSync for different
architectures or when the executable is outdated. This allows easy installation
on remote machines: Zip McSync root folder, send it over the network and then
run mcsync.sh as always.