SSH related information is now stored in plug for easy access (session member)
The reason for this is that we need a place to store the path to the SSH session.
This, together with the machine name and the user name will allow us to start
file transfers using scp or rsync which both know how to use the SSH session
multiplexing feature.