Changes the whole history handling back to a grafts based approach
In order to simplify handling, the scanning and history keeping now
is based only on grafts. Every scan request will lead to a full graft rescan as this is the easiest to implement. Otherwise more
complex handling also of the history files is needed. Further,
a naming scheme for the history files has been implemented.